backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
passlib[bcrypt]
PyPDF2
requests
openai
fpdf
python-multipart
pydantic
email-validator
